Update README.md
Project Title : "SimpeeDine" An Intelligent Food Demand 
Forecasting Management System in Restaurants 
According to the Food and Agriculture Organization of the United Nations (FAO) 75% of the food 
produced all around the world (only for human consumption) just over billion tons per year is 
wasted without consuming. At the same time, around the world about 900 million people are 
starving and suffering from malnutrition and hunger. (Bronte; 2021). 
One of the most important parts of maximizing revenue and also satisfying customers is the 
perfect amount of food. When demand is lower than production, assets and resources will be 
wasted, but when the demand is higher than the supply, there will be a lack of consumer 
satisfaction. 
This food wastage situation is a real-life challenge for resource optimization, which can be 
efficiently overcome by pattern recognition, and data mining techniques. Those demand 
forecasting systems are developed to forecast the future demand for a specific product or service. 
This is an end result of the project to build such a demand forecasting system. 
The proposed solution, SimpeeDine will be built by using the historical data of previous sales, 
promotions and discount data to get predictive output of the demand for a particular food item for 
the new coming week. The system SimpeeDine is mainly relied on regression and machine 
learning models that is trained and tested by applying a Voting Regressor. Some other algorithms 
have also been used for regression. A graphical user interface can be used for the end user to 
interact with the system. Forecasting is made by analyzing the inputs which have been already 
gathered. The demand of the particular food items will be predicted separately (one by one) .The 
results of the output can be displayed in the system.